% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_criteria.R
\name{ar_information_criteria}
\alias{ar_information_criteria}
\title{Model selection criteria for AR models}
\usage{
ar_information_criteria(y, order_limit = 6)
}
\arguments{
\item{y}{time series}

\item{order_limit}{limit to the AR model order (integer)}
}
\value{
a list of the optimal order found for each criterion and the criteria
  values
}
\description{
This calculates the statistics and optimal model orders for AR time series
models
}
\examples{
set.seed(1)
y <- arima.sim(list(ar = c(.35,.1)), n = 1200)
ar_information_criteria(y,10)
}
