% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_diags.R
\name{basic_diags}
\alias{basic_diags}
\title{Utility to check some basic diagnostics}
\usage{
basic_diags(fit)
}
\arguments{
\item{fit}{Fitted stan object}
}
\description{
Utility to check some basic diagnostics
}
