% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distpar.R
\name{calc_distpar}
\alias{calc_distpar}
\title{Utility to turn mean/variance into gamma or NB parameters}
\usage{
calc_distpar(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{var}{variance}
}
\description{
Utility to turn mean/variance into gamma or NB parameters
}
