% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stz.R
\name{check_stz}
\alias{check_stz}
\title{See how closely batches of parameters are summing to 0}
\usage{
check_stz(fit, pars = "theta", excl_pars = "sigma", print = TRUE, ret = FALSE)
}
\arguments{
\item{fit}{Fitted stan object}

\item{pars}{Pattern matching which parameters to explore, defaults to everything named theta}

\item{excl_pars}{Pattern for which parameters to exclude. Takes precedence over pars. Defaults to sigma}

\item{print}{Whether to print the density plots}

\item{ret}{Whether to return the data comparing parameters (TRUE) or the plot (FALSE)}
}
\description{
See how closely batches of parameters are summing to 0
}
