% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_pars.R
\name{compare_pars}
\alias{compare_pars}
\title{Compare parameter posteriors from fitted models}
\usage{
compare_pars(
  fitlist,
  pars = "sigma",
  excl_pars = "raw|^lp__$",
  print = TRUE,
  ret = FALSE
)
}
\arguments{
\item{fitlist}{Named list of fitted stan objects}

\item{pars}{Pattern matching which parameters to explore, defaults to 'sigma'}

\item{excl_pars}{Pattern for which parameters to exclude. Takes precedence over pars. Defaults to raw and lp__}

\item{print}{Whether to print the density plots}

\item{ret}{Whether to return the data comparing parameters (TRUE) or the plot (FALSE)}
}
\description{
Compare parameter posteriors from fitted models
}
