% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_prior.R
\name{compare_prior}
\alias{compare_prior}
\title{Utility to compare a batch of parameters to their prior}
\usage{
compare_prior(
  fit,
  pars = "sigma",
  excl_pars = "raw|^lp__$",
  prior = abs(rnorm(1e+05, 0, 1))
)
}
\arguments{
\item{fit}{Fitted stan object}

\item{pars}{Pattern for which parameters to include, defaults to all things sigma}

\item{excl_pars}{Pattern for which parameters to exclude (takes precedence over pars). Defaults to excluding anything with 'raw' and lp__}

\item{prior}{A vector with draws from the prior; defaults to N+(0,1)}
}
\description{
Utility to compare a batch of parameters to their prior
}
