% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_all.R
\name{compile_all}
\alias{compile_all}
\title{Wrapper to compile_model which runs for all models in a folder}
\usage{
compile_all(models = NULL, model_folder = "models")
}
\arguments{
\item{models}{optional regex to determine which models to compile. Defaults to NULL, compiling all}

\item{model_folder}{Path to models, defaults to "models"}
}
\description{
Wrapper to compile_model which runs for all models in a folder
}
