% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dan.R
\name{gamma_params}
\alias{gamma_params}
\title{Gamma Parameters}
\usage{
gamma_params(mean, var)
}
\arguments{
\item{mean}{Desired mean of the distribution}

\item{var}{Desired variance of the distribution}
}
\description{
This function gives  appropriate shape and rate parameters for a gamma distribution given a desired mean and variance
}
\examples{
> gamma_params(mean = 5, var = 3)
[1] 8.333333 1.666667
> mean(rgamma(1000,8.333333,1.666667))
[1] 4.937651
> var(rgamma(1000,8.333333,1.666667))
[1] 3.172861
}
\keyword{Gamma,}
\keyword{hyperparameters}

