% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_smry.R
\name{get_smry}
\alias{get_smry}
\title{Wrapper for rstan's summary that returns a prettier tibble and uses a regex to filter parameters}
\usage{
get_smry(
  fit,
  pars = ".",
  excl_pars = "raw|^lp__$",
  excl_probs = TRUE,
  excl_se = TRUE,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975)
)
}
\arguments{
\item{fit}{Fitted stan object}

\item{pars}{Pattern for which parameters to include}

\item{excl_pars}{Pattern for which parameters to exclude (takes precedence over pars). Defaults to excluding anything with 'raw' and lp__}

\item{excl_probs}{Whether to exclude the posterior probability calculates, defaults to TRUE}

\item{excl_se}{Whether to exclude the SE of the posterior mean, defaults to TRUE}

\item{probs}{Which posterior probabilities to include. Only does anything if excl_probs is FALSE}
}
\description{
Wrapper for rstan's summary that returns a prettier tibble and uses a regex to filter parameters
}
