% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dan.R
\name{inv_gamma_params}
\alias{inv_gamma_params}
\title{Inverse Gamma Parameters}
\usage{
inv_gamma_params(mean, var)
}
\arguments{
\item{mean}{Desired mean of the distribution}

\item{var}{Desired variance of the distribution}
}
\description{
This function gives appropriate shape and rate parameters for an inverse gamma distribution when you give a desired mean and variance for the distribution.
}
\examples{
inv_gamma_params(mean = 2,var = 1)
[1]  6 10
mean(1/rgamma(1000,6,10))
[1] 2.027287
var(1/rgamma(1000,6,10))
[1] 0.9945234
}
\keyword{Gamma,}
\keyword{Inverse}
\keyword{hyperparameters}

