% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{invlogit}
\alias{invlogit}
\title{Utility to calculate the probability form a logit score}
\usage{
invlogit(x)
}
\arguments{
\item{x}{vector of logits}
}
\description{
Utility to calculate the probability form a logit score
}
