% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify_draws.R
\name{lisitfy_draws}
\alias{lisitfy_draws}
\title{Convert draws_array from cmdstan to rstan-type named list of parameters, with vectors/matrices/arrays unflattened}
\usage{
lisitfy_draws(fit, draws, flatten_chains = FALSE)
}
\arguments{
\item{fit}{cmdstan fit object}

\item{draws}{draws from cmdstan}

\item{flatten_chains}{whether to leave iters & chains as separate dimensions or combine}
}
\description{
Convert draws_array from cmdstan to rstan-type named list of parameters, with vectors/matrices/arrays unflattened
}
