% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listify_smry.R
\name{listify_smry}
\alias{listify_smry}
\title{Convert stan summary (e.g. from get_smry) to named list of posterior means}
\usage{
listify_smry(smry)
}
\arguments{
\item{smry}{object to convert}
}
\description{
Convert stan summary (e.g. from get_smry) to named list of posterior means
}
