% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pkg_fns.R
\name{load_pkg_fns}
\alias{load_pkg_fns}
\title{Load visible and invisible package functions}
\usage{
load_pkg_fns(pkg_name)
}
\arguments{
\item{pkg_name}{The string of the package whose functions you would like
to load.}
}
\description{
This tends to be useful when debugging
}
\examples{
\dontrun{load_pkg_fns(BayesfMRI)}
}
