% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Utility to calculate a logit score from a probability}
\usage{
logit(p)
}
\arguments{
\item{p}{vector of probabilities}
}
\description{
Utility to calculate a logit score from a probability
}
