% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree_plot.R
\name{make_tree_plot}
\alias{make_tree_plot}
\title{Utility to make basic CART tree}
\usage{
make_tree_plot(y, x, cp = 0.1, pfx = "y", ...)
}
\arguments{
\item{y}{y}

\item{x}{x}

\item{cp}{CART's complexity parameter thingy}

\item{pfx}{Prefix for plot}

\item{...}{additional arguments passed to rpart::rpart}
}
\description{
Utility to make basic CART tree
}
