% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Wrapper for rstan's monitor that returns a prettier tibble}
\usage{
monitor(chains, warmup = 0, probs = 0.5, extra_cols = NULL)
}
\arguments{
\item{chains}{Stan-style array of iters*chains*pars}

\item{warmup}{what number of draws is excluded as warmup. Default assumes 0 for sampling only}

\item{probs}{What probabilities to calculate}

\item{extra_cols}{Which additional columns, like mcmc SE, to keep. If you want probs then you need to specify them here too because Iam lazy}
}
\description{
Wrapper for rstan's monitor that returns a prettier tibble
}
