% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vtt.R
\name{parse_vtt}
\alias{parse_vtt}
\title{Utility to approximate one distribution with another}
\usage{
parse_vtt(file, smry = FALSE)
}
\arguments{
\item{file}{vtt file}

\item{smry}{whether to return raw data (FALSE, default) or a summary of the data (TRUE)}
}
\description{
Utility to approximate one distribution with another
}
