% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastex.R
\name{pastex}
\alias{pastex}
\title{Crosspaste two vectors of strings}
\usage{
pastex(x, y, sep = "", collapse = NULL)
}
\arguments{
\item{x}{First string vector}

\item{y}{Decond string vector}

\item{sep}{Separator between values of x and y}

\item{collapse}{If NULL, the default, then returns a vector of each x-y pair;
if not null returns a single string, with each x-y pair separated by the value of collapse}
}
\description{
Crosspaste two vectors of strings
}
