% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_corr.R
\name{post_corr}
\alias{post_corr}
\title{Utility to create a pretty-ish heatmap of posterior correlations}
\usage{
post_corr(
  fit,
  pars = "sigma",
  excl_pars = "raw|lp__",
  label = TRUE,
  label_round = 2,
  ...
)
}
\arguments{
\item{fit}{Fitted stan object}

\item{pars}{Pattern for which parameters to include}

\item{excl_pars}{Pattern for which parameters to exclude (takes precedence over pars). Defaults to excluding anything with 'raw' and lp__}

\item{label}{Whether to print corrs in the matrix, defaults to TRUE}

\item{label_round}{Number of digits for correlation labels, defaults to 2}

\item{...}{options passed to GGally::ggcorr()}
}
\description{
Utility to create a pretty-ish heatmap of posterior correlations
}
