% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_approx.R
\name{qq_approx}
\alias{qq_approx}
\title{Utility to approximate one distribution with another}
\usage{
qq_approx(posterior, dist = "norm", n = 1000, distance = "sq")
}
\arguments{
\item{posterior}{posterior draws you wish to approximate}

\item{dist}{Which distribution to approximate with}

\item{n}{how many quantiles to use}
}
\description{
Utility to approximate one distribution with another
}
