% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_pars.R
\name{regex_pars}
\alias{regex_pars}
\title{Utility to get a list of parameters via regex}
\usage{
regex_pars(fit, pars = ".", excl_pars = NULL)
}
\arguments{
\item{fit}{The fitted stan object}

\item{pars}{Pattern for which parameters to use, defaults to '.' to select everything}

\item{excl_pars}{Pattern for which parameters to exclude. Takes precedence over pars}
}
\description{
Utility to get a list of parameters via regex
}
