% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regextract.R
\name{regextract}
\alias{regextract}
\title{Wrapper around extract to allow regular expressions}
\usage{
regextract(fit, regex = ".", as_df = TRUE)
}
\arguments{
\item{fit}{The fitted stan object}

\item{regex}{Pattern, defaults to '.' to select everything}

\item{as_df}{whether to return the default list thing like stan, or a tibble}
}
\description{
Wrapper around extract to allow regular expressions
}
