% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dan.R
\name{rmvn}
\alias{rmvn}
\title{Multivariate Normal Draws}
\usage{
rmvn(n, mu = 0, V = matrix(1))
}
\arguments{
\item{n}{The number of draws to make from the multivariate normal distribution. Must be a positive integer.}

\item{mu}{The mean vector of length p}

\item{V}{The $p\times p$ covariance matrix}
}
\description{
This function draws from a multivariate normal distribution of dimension p more efficiently than the rmvnorm function in the mvtnorm package.
}
\examples{
rmvn(5,c(2,2),matrix(c(2,-1,-1,5),2,2))
          [,1]      [,2]      [,3]     [,4]     [,5]
[1,] 0.2962443 3.3602800 0.9445853 2.869058 1.860385
[2,] 7.5907838 0.4752729 1.7318177 1.648304 3.482479
}
\keyword{Gaussian,}
\keyword{Normal}
