% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_diff.R
\name{var_diff}
\alias{var_diff}
\title{Compare variable names in two different datasets}
\usage{
var_diff(x, y, quiet = FALSE, ret = FALSE)
}
\arguments{
\item{x}{First dataset}

\item{y}{Second dataset}

\item{quiet}{Whether to print info about overlap}

\item{ret}{Whether to return list with vector of overlapping vars}
}
\description{
Compare variable names in two different datasets
}
