% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{Wrapper to compile model and save as RDS. Only recompiles if stan file has changed}
\usage{
compile_model(model, model_folder = "models")
}
\arguments{
\item{model}{Name of model to compile. Exclude the ".stan" ending}

\item{model_folder}{Path to model, defaults to "models"}
}
\description{
Wrapper to compile model and save as RDS. Only recompiles if stan file has changed
}
