% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_criteria.R
\name{gelfand_and_ghosh}
\alias{gelfand_and_ghosh}
\title{Gelfand and Ghosh Posterior Predictive Loss Criterion Function}
\usage{
gelfand_and_ghosh(k, preds, obs)
}
\arguments{
\item{k}{This is the weight parameter that decides how much goodness-of-fit
matters in your loss function.}

\item{preds}{This is an n by s matrix of predictions}

\item{obs}{This is an n-dimensional observation vector}
}
\description{
This function implements the Gelfand and Ghosh predictive loss criterion as
proposed in 1998 by Alan Gelfand and Sujit Ghosh.
}
\references{
Gelfand, Alan E., and Sujit K. Ghosh. "Model choice: a minimum
  posterior predictive loss approach." Biometrika 85.1 (1998): 1-11.
}
\keyword{Bayesian}
\keyword{Comparisons,}
\keyword{Model}
