% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regextract.R
\name{regextract}
\alias{regextract}
\title{Wrapper around extract to allow regular expressions}
\usage{
regextract(fit, pars = ".", excl_pars = "raw|^lp__$", as_df = TRUE)
}
\arguments{
\item{fit}{The fitted stan object}

\item{pars}{Pattern for which parameters to use, defaults to '.' to select everything}

\item{excl_pars}{Pattern for which parameters to exclude. Takes precedence over pars. Defaults to raw and lp__}

\item{as_df}{Whether to return the default list thing like stan, or a tibble}
}
\description{
Wrapper around extract to allow regular expressions
}
